% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outfiles.R
\name{get_outfiles}
\alias{get_outfiles}
\title{get_outfiles}
\usage{
get_outfiles(dir = ".")
}
\arguments{
\item{dir}{Directory into which files ought to be stored. Default current
work directory.}
}
\value{
deposits six *.dist StereoGene output files into the selected
directory
}
\description{
Copy files necessary to complete the vignette onto the local
machine in cases where Stereogene, CapR, or bedtools are not available.
}
\examples{
## pull example StereoGene output files
get_outfiles()

}
